% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Diabetes-data.R
\docType{data}
\name{Diabetes}
\alias{Diabetes}
\title{Dataset of Diabetes}
\format{
An object of class \code{"data.frame"}
}
\source{
\href{https://data.mendeley.com/datasets/wj9rwkp9c2/1}{Link}
}
\usage{
data(Diabetes)
}
\description{
Diabetes, also known as diabetes mellitus, is a group of metabolic
disorders characterized by a high blood sugar level (hyperglycemia) over
a prolonged period of time. Symptoms often include frequent urination,
increased thirst and increased appetite. If left untreated, diabetes can
cause many health complications. Acute complications can include diabetic
ketoacidosis, hyperosmolar hyperglycemic state, or death. Serious long-term
complications include cardiovascular disease, stroke, chronic kidney disease,
foot ulcers, damage to the nerves, damage to the eyes, and cognitive impairment.
}
\examples{
data(Diabetes)
}
\keyword{dataset}
