% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data_final}
\alias{get_data_final}
\title{Get data}
\usage{
get_data_final(x)
}
\arguments{
\item{x}{the path of the data}
}
\value{
the data set
}
\description{
This function is used for reading data
}
\keyword{data}
\keyword{read}
