% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_histogram.R
\name{get_histogram_final}
\alias{get_histogram_final}
\title{Get a histogram}
\usage{
get_histogram_final(df, col)
}
\arguments{
\item{df}{the data set}

\item{col}{column name}
}
\value{
the histogram
}
\description{
This function is used to get a histogram
}
\examples{
get_histogram_final(cars, "speed")
}
\keyword{histogram}
