% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lm.R
\name{get_lm_final}
\alias{get_lm_final}
\title{Get a simple linear regression model}
\usage{
get_lm_final(df, col_y, col_x)
}
\arguments{
\item{df}{the data set}

\item{col_y}{column name}

\item{col_x}{column name}
}
\value{
simple linear regression results
}
\description{
This function is used to do simple linear regression
}
\examples{
get_lm_final(cars, "speed", "dist")
}
\keyword{linear}
\keyword{regression}
\keyword{simple}
