% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_summarize.R
\name{get_summarize_final}
\alias{get_summarize_final}
\title{Get data summaries}
\usage{
get_summarize_final(df, col)
}
\arguments{
\item{df}{the data set}

\item{col}{column name}
}
\value{
data_summary
}
\description{
This function is used to get data summaries
}
\examples{
get_summarize_final(cars, "speed")

}
\keyword{deviation,}
\keyword{maximum,}
\keyword{mean,}
\keyword{median,}
\keyword{minimum}
\keyword{standard}
